+++ protected/humhub/modules/ui/form/widgets/views/colorPickerField.php
@@ -3,6 +3,12 @@
 use yii\helpers\Html;
 
 humhub\assets\HumHubColorPickerAsset::register($this);
+
+/* @var $model \yii\base\Model*/
+/* @var $field string */
+/* @var $inputId string */
+/* @var $container string */
+
 ?>
 
 <?= Html::activeTextInput($model, $field, ['class' => 'form-control', 'id' => $inputId, 'value' => $model->$field, 'style' => 'display:none']); ?>